feat(Home): homepage layout implemented containing table and chart
The homepage layout was implemented following the provided template (https://bit.ly/3Cle1Bx), the 'ApexCharts' lib was used to create the donut chart and some libs from MUI were used to implement the inputs fields, the 'send' button and the table. The 'styled-components'  lib was added to, for styling.